--- src/KakaoMap.tsx
@@ -1,70 +1,0 @@
-// 카카오 지도 맵 관련 코드
-import React, { useEffect, useRef } from 'react';
-
-declare global {
-  interface Window {
-    kakao: any;
-  }
-}
-
-const KakaoMap = () => {
-  const mapRef = useRef<any>(null);
-  const markerRef = useRef<any>(null);
-
-  useEffect(() => {
-    const script = document.createElement('script');
-    script.src = `https://dapi.kakao.com/v2/maps/sdk.js?appkey=d735989af4ed4e049847dd3e7d4fba63&autoload=false`; // autoload=false 추가
-    script.async = true;
-    document.head.appendChild(script);
-
-    script.onload = () => {
-      window.kakao.maps.load(() => {
-        console.log('카카오 지도가 로드되었습니다.');
-        const container = document.getElementById('map');
-        if (!container) {
-          console.error('표시할 영역을 찾을 수 없습니다.');
-          return;
-        }
-
-        const options = {
-          center: new window.kakao.maps.LatLng(33.450701, 126.570667), 
-          level: 3, 
-        };
-
-        const map = new window.kakao.maps.Map(container, options);
-        mapRef.current = map;
-
-        const marker = new window.kakao.maps.Marker({
-          position: map.getCenter(), 
-          map: map,
-        });
-        markerRef.current = marker;
-
-        window.kakao.maps.event.addListener(map, 'center_changed', () => {
-          const center = map.getCenter(); 
-          marker.setPosition(center); 
-        });
-
-        console.log('지도에 마커가 추가되었습니다.', map.getCenter());
-      });
-    };
-
-    script.onerror = () => {
-      console.error('카카오 지도를 로드하지 못했습니다.');
-    };
-
-    return () => {
-      script.remove(); 
-    };
-  }, []);
-
-  return (
-    <div>
-      <h1>Kakao 지도 예제 - 지도 중앙에 마커 고정</h1>
-      {/* 카카오 지도 */}
-      <div id="map" style={{ width: '500px', height: '400px' }}></div>
-    </div>
-  );
-};
-
-export default KakaoMap;
